+++ Examples/ConsoleApplication/ConsoleApplication/ConsoleApplication.cpp
@@ -1,0 +1,27 @@
+#include <cstdio>
+
+/* 
+
+	A console application created through the BuildBridge Project Wizard.
+	Press F5 to assemble the app and launch it.
+	Press ctrl + shift + B to simply assemble the app.
+	---------------------------------------------------------------------
+	There are three buttons on the "Build Bridge -> Projects" panel.
+	1 (left): Build the application (analogous to ctrl + shift + B).
+	2: Launch the application with Debugger.
+	3: Launch the application without Debugger.
+	---------------------------------------------------------------------
+
+*/
+
+
+/* 
+	Default ConsoleApplication entry point
+*/
+int main()
+{
+	printf("It's simple Console Application!\n");
+	printf("It's created via BuildBridge Project Wizard!\n");
+	printf("Have Fun\n");
+	return 0;
+}