+++ Examples/SharedLibrary/TestLIB/TestLIB.h
@@ -1,0 +1,16 @@
+
+class TestPrinter
+{
+private:
+    
+public:
+    TestPrinter();
+    ~TestPrinter();
+
+    /* 
+        Test function 
+        Just print message in console
+        Using 'printf'
+    */
+    void PrintMessage(const char* message);
+};