+++ media/WebView/CreateProject/PrepareToCreate.js
@@ -1,6 +1,8 @@
 
-
-function SendDataToVsCode(projectData) {
+/* 
+    Prepare project data and send in vs code
+*/
+function SendDataToVsCode(projectData, projectType) {
     console.log('vs code received data: ', projectData);
 
     const data = {
@@ -8,7 +10,7 @@ function SendDataToVsCode(projectData) {
         platform: projectData.projectPlatform,
         buildSystem: 'MSBuild',
         configuration: projectData.projectConfiguration,
-        internalCommand: 'Create_Console_Application'
+        internalCommand: projectType
     };
 
     // Notify vs code that a project needs to be created