+++ src/Utils/ExtensionVersion.ts
@@ -1,0 +1,7 @@
+
+
+/* Helper function for get extension version */
+export function ReturnCurrentExtensionVersion()
+{
+    return 'v0.1.3';
+}