+++ CHANGELOG.md
@@ -1,9 +1,34 @@
-# Change Log
+# 🚀 BuildBridge Changelog
 
-All notable changes to the "buildbridge" extension will be documented in this file.
+<!-- Version 0.1.3 -->
+## 🔥 Version 0.1.3 — Project Types Expansion
+📅 Release Date: 18.03.2026
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+| Issue | Description | Note |
+|-------|-------------|------|
+| ❌ Wizard missing .lib/.dll       | ✅ Now supports all 3 templates!    | 📌 exe, lib, dll |
+| ❌ Tree view delete/create broken | ✅ Full file operations in projects | 📌 Files sync with .vcxproj |
+| ❌ Old Version of Extension on the Github | ✅ Updated Github page | 📌 Fresh files in the github |
+| ❌ There were no buttons for building and launching the project. | ✅ Added buttons: 'Build project', 'Run(Debug, without Debugger)' | 📌 Added new buttons in Project Expolorer |
+| ❌ No Examples | ✅ Added examples for each project type | 📌 Examples only on the Github!(https://github.com/SergeyNaumenkov/BuildBridge) |
+| - | ✅ Added New View in Project Explorer | 📌 WebPanel with HTML content about extension |
 
-## [Unreleased]
+---
 
-- Initial release
+<!-- Version 0.1.2 -->
+## ⚡ Version 0.1.2 — Dependencies Fix
+📅 Release Date: 16.03.2026
+
+| Issue | Description | Note |
+|-------|-------------|------|
+| ❌ Project Dependencies not loading | ✅ Dependencies now work properly | 📌 Uses `xmldom` |
+
+---
+
+<!-- Version 0.1.1 -->
+## 📄 Version 0.1.1 — Documentation
+📅 Release Date: 15.03.2026
+
+| Issue | Description | Note |
+|-------|-------------|------|
+| - | 🔥 Just edit README.md file | - |